docs: Implement new Sidebar and Folder Structure
* docs: Implement new Sidebar and Folder Structure
