+++ .storybook/preview.tsx
@@ -4,7 +4,7 @@ import { dark, light } from "./frontifyTheme";
 export const parameters = {
     options: {
         storySort: {
-            order: ["Foundation", "Components"],
+            order: ["Tokens", "Layout", "Typography", "Components"],
         },
     },
     darkMode: {
