+++ vite.config.ts
@@ -12,7 +12,9 @@ export const alias = {
     "@components": resolve(__dirname, "./src/components"),
     "@foundation": resolve(__dirname, "./src/foundation"),
     "@hooks": resolve(__dirname, "./src/hooks"),
+    "@layout": resolve(__dirname, "./src/layout"),
     "@utilities": resolve(__dirname, "./src/utilities"),
+    "@typography": resolve(__dirname, "./src/typography"),
 };
 
 export const globals = {
